requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
langchain-groq
gTTS
pydub